Stop putting packages in Packages/deb with kirb/theos
Previously THEOS_BUILD_DIR was set in the makefile, so theos would make
the packages in Packages/ and keep everything tidy. The new theos used,
kirb/theos, automatically does this with the deb/ folder, so that line
is now redundant and actually put packages in Packages/deb/.

Leave the Packages/ in the .gitignore for backwards compatability with
previous forks of theos and/or the Makefile.